+++ Sources/AIProxy/AIProxy.swift
@@ -975,6 +975,37 @@ public enum AIProxy {
         )
     }
 
+    /// Returns the authentication and routing metadata needed to send an existing gRPC call through AIProxy.
+    ///
+    /// Add the returned key-value pairs to the metadata on your gRPC call. Your gRPC client remains responsible
+    /// for the transport, method path, serialization, and response handling.
+    ///
+    /// - Parameters:
+    ///   - partialKey: Your partial key from the AIProxy developer dashboard.
+    ///   - serviceURL: Your service URL from the AIProxy developer dashboard.
+    ///   - clientID: An optional client ID used to attribute requests to a specific user or device.
+    /// - Returns: The metadata fields expected by the AIProxy gRPC endpoint.
+    nonisolated public static func grpcMetadata(
+        partialKey: String,
+        serviceURL: String,
+        clientID: String? = nil
+    ) async throws -> [String: String] {
+        let identifiers = try AIProxyUtils.serviceIdentifiers(from: serviceURL)
+        let request = try await self.request(
+            partialKey: partialKey,
+            serviceURL: serviceURL,
+            clientID: clientID,
+            proxyPath: "/"
+        )
+
+        var metadata = (request.allHTTPHeaderFields ?? [:]).reduce(into: [String: String]()) {
+            $0[$1.key.lowercased()] = $1.value
+        }
+        metadata["aiproxy-project"] = identifiers.project
+        metadata["aiproxy-service"] = identifiers.service
+        return metadata.filter { $0.key.hasPrefix("aiproxy-") }
+    }
+
 #if canImport(AppKit) && !targetEnvironment(macCatalyst)
     nonisolated public static func encodeImageAsJpeg(
         image: NSImage,
